Fix: Break out constants and find activity function to external files
